requirements.txt
asgiref==3.8.1
attrs==23.2.0
distlib==0.3.8
Django==4.2
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.13.3
inflection==0.5.1
javascript==1!1.2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
whitenoise==6.7.0
